move getAncetsorArray somewhere more sensible
